Move AST supports into own compile units
