+++ Makefile.conf
@@ -8,6 +8,7 @@
 SOURCES = \
 	ast.cpp \
 	ast_values.cpp \
+	ast_supports.cpp \
 	ast_sel_cmp.cpp \
 	ast_sel_unify.cpp \
 	ast_selectors.cpp \
