+++ win/libsass.targets
@@ -13,6 +13,7 @@
     <ClInclude Include="$(LIBSASS_HEADERS_DIR)\ast.hpp" />
     <ClInclude Include="$(LIBSASS_HEADERS_DIR)\ast_values.hpp" />
     <ClInclude Include="$(LIBSASS_HEADERS_DIR)\ast_selectors.hpp" />
+    <ClInclude Include="$(LIBSASS_HEADERS_DIR)\ast_supports.hpp" />
     <ClInclude Include="$(LIBSASS_HEADERS_DIR)\ast_def_macros.hpp" />
     <ClInclude Include="$(LIBSASS_HEADERS_DIR)\memory\SharedPtr.hpp" />
     <ClInclude Include="$(LIBSASS_HEADERS_DIR)\ast_factory.hpp" />
@@ -80,6 +81,7 @@
   <ItemGroup Label="LibSass Sources">
     <ClCompile Include="$(LIBSASS_SRC_DIR)\ast.cpp" />
     <ClCompile Include="$(LIBSASS_SRC_DIR)\ast_values.cpp" />
+    <ClCompile Include="$(LIBSASS_SRC_DIR)\ast_supports.cpp" />
     <ClCompile Include="$(LIBSASS_SRC_DIR)\ast_sel_cmp.cpp" />
     <ClCompile Include="$(LIBSASS_SRC_DIR)\ast_sel_unify.cpp" />
     <ClCompile Include="$(LIBSASS_SRC_DIR)\ast_selectors.cpp" />
